pyproject.toml
[tool.poetry]
name = "hikari-bot"
version = "0.1.0"
description = ""
authors = ["Nishant Sapkota"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
python-dotenv = "^0.19.0"
aiohttp = "^3.7.4"
aerich = "0.5.3"
tortoise-orm = "0.17.4"
asyncpg = "^0.24.0"
hikari = {git = "https://github.com/hikari-py/hikari"}
hikari-tanjun = {git = "https://github.com/FasterSpeeding/Tanjun"}
APScheduler = "^3.8.0"
bs4 = "^0.0.1"
psutil = "^5.8.0"
hikari-lightbulb = {git = "https://github.com/tandemdude/hikari-lightbulb"}
hikari-yuyo = {git= "https://github.com/FasterSpeeding/Yuyo"}
lavasnek-rs = {git = "https://github.com/vicky5124/lavasnek_rs"}

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
